Refactor: 의존성 리펙터링 - order 와 table 간 연관관계 제거 - menu 와 product 패키지 병합
